Make AST nodes minimalistic and extensible
AST nodes from this commit on will only provide the absolute
required fields out of the box (for example a binary has three
nodes only, lhs, rhs and operator). At the same time, we can
extend the fields of AST nodes via a so-called attributes which
for the user is exactly like fields but it is in reality built
using a key-value map.

This way, we make it possible (and easy) to extend AST nodes with
new properties (fields).