Change decls to List not Set
While it makes more sense to have them as Set, but when we
first define symbols they all look the same at the symbolassigner
stage, and the result will be they won't be added to the set of
decls.

For examples at symbol assigners the following two ``test'' methods
have symbols that are equal to each other.

```
class A {
  void test(boolean p) {}
  void test(int p) {}
}
```

The reason behind this is that types are not assigned at this stage,
and overloaded methdos rely on their signature to standout.