Make namer name top-levels & typer name the rest
Now, namer names the top-level only, and leaves the others to typer.
(Right now, for top-level idents typer repeats everything namer does,
and concludes the same results as nmaer, but this is not a big deal,
a simple if-statement should fix it.)

Next would be, make symbolassigner assign symbols for top-level fields
only and make typer assign symbols for local variables and parameters.
This will be the last step, until getting rid of flow information of
typer.