Add family macro annotation
This macro annotation uses an if-else based family method dispatcher, in
contrast to the while-loop or orElse method based ones.

The performance of this is on par with the while-loop, but the code can
get much messier, as we cannot open macro annotation parameters unless
they are literals (not even contant experssions!!).